requirements.txt
Django==1.8.2
django-ajaximage==0.2.8
django-braces==1.8.0
django-crispy-forms==1.4.0
django-import-export==0.2.7
tablib==0.10.0
